pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "flask>=3.1.3",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.1",
    "markdown>=3.10.2",
    "python-dotenv>=1.2.1",
    "unstructured>=0.20.8",
]
